third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-invalid-cycles.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    width: 100px;
    height: 100px;
    background-color: green;
    --a1: { @apply --a2; background-color: red; };
    --a2: { @apply --a1; background-color: red; };
    @apply --a1;
    @apply --a2;
    --b1: { @apply --b2; background-color: red; };
    --b2: var(--b3);
    --b3: { var(--b1, background-color: red); };
    @apply --b1;
    @apply --b2;
    @apply --b3;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).getPropertyValue("--a1"), "");
    assert_equals(getComputedStyle(e).getPropertyValue("--a2"), "");
    assert_equals(getComputedStyle(e).getPropertyValue("--b1"), "");
    assert_equals(getComputedStyle(e).getPropertyValue("--b2"), "");
    assert_equals(getComputedStyle(e).getPropertyValue("--b3"), "");
}, "Variable cycle detection works with @apply");
</script>
